// src/app/shared/models/booking.model.ts
export class Booking {
  bookingId?: string;
  bookingStatus?: string;
  dateSlot?: Date;
  dayWithDate?: string;
  doctorId?: string;
  creationDate?: Date;
  doctorMobile?: string;
  doctorName?: string;
  doctorSpeciality: string;
  modificationDate?: Date;
  id?: string;
  email?: string;
  patientMobileNumber?: string;
  patientAge?: number;
  patienCity?: number;
  patientGender?: string;
  city?: string;
  patientAddress?: string;
  patientName?: string;
  patientState?: string;
  state?: string;
  symptoms?: string;
  timeSlot?: string;
  timing?: string;
  userId?: string;
}